# api/internal/docs/combined_swagger.yaml
openapi: 3.0.0
info:
  title: Pseudo App API
  description: Complete API documentation for Pseudo App endpoints including authentication, status, and health checks
  version: 1.0.0
  contact:
    name: API Support
    email: support@pseudo-app.com

servers:
  - url: http://localhost:8000/api/v1
    description: Development server

paths:
  # Authentication endpoints
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: "John Doe"
                  description: User's full name
                email:
                  type: string
                  format: email
                  example: "john.doe@example.com"
                  description: User's email address
                password:
                  type: string
                  format: password
                  example: "password123"
                  description: User's password (minimum 6 characters)
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/UserResponse'
                      access_token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      refresh_token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/signin:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "john.doe@example.com"
                  description: User's email address
                password:
                  type: string
                  format: password
                  example: "password123"
                  description: User's password
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/UserResponse'
                      access_token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      refresh_token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get current authenticated user information
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "success"
                  data:
                    $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh-token:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Generate new access token using refresh token
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  description: Valid refresh token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      refresh_token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Unauthorized - Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Logout user and invalidate tokens
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "success"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Status and Health endpoints
  /status:
    get:
      summary: Get application status
      description: Returns the current status of the application
      tags:
        - Status
      responses:
        '200':
          description: Application status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "success"
              examples:
                success:
                  summary: Successful status response
                  value:
                    message: "success"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      summary: Health check endpoint
      description: Checks the health status of the application and its dependencies
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  message:
                    type: string
                    example: "Service is running"
                  database:
                    type: string
                    example: "connected"
              examples:
                healthy:
                  summary: Healthy service response
                  value:
                    status: "healthy"
                    message: "Service is running"
                    database: "connected"
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "unhealthy"
                  message:
                    type: string
                    example: "Database connection failed"
                  error:
                    type: string
                    example: "connection refused"
              examples:
                unhealthy:
                  summary: Unhealthy service response
                  value:
                    status: "unhealthy"
                    message: "Database connection failed"
                    error: "connection refused"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
          description: User ID
        name:
          type: string
          example: "John Doe"
          description: User's full name
        email:
          type: string
          format: email
          example: "john.doe@example.com"
          description: User's email address
        created_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
          description: User creation timestamp
        updated_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
          description: User last update timestamp

    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: "failed"
        error:
          type: string
          example: "Validation failed"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Email is required"

    UnauthorizedError:
      type: object
      properties:
        message:
          type: string
          example: "failed"
        error:
          type: string
          example: "Unauthorized access"

    StatusResponse:
      type: object
      properties:
        message:
          type: string
          description: Status message
      required:
        - message

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          description: Health status
          enum: [healthy, unhealthy]
        message:
          type: string
          description: Health message
        database:
          type: string
          description: Database connection status
      required:
        - status
        - message

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error description
      required:
        - error

tags:
  - name: Authentication
    description: User authentication and authorization operations
  - name: Status
    description: Application status operations
  - name: Health
    description: Health check operations